refactor: api 호출 파일 위치 이동
- /service 디렉토리 생성 후 이동
