+++ client/components/Artwork/Tiles/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import styled from '@emotion/styled';
 
-import { getAllArtworks } from 'utils/networking';
+import { getAllArtworks } from 'service/networking';
 import ImageTile from '../ImageTile';
 import { Artwork } from 'interfaces';
