+++ client/components/Exhibition/ArtworkSelector/index.tsx
@@ -1,11 +1,8 @@
 import React, { useEffect, useState } from 'react';
 
 import { Artwork } from 'interfaces';
-import {
-    useSelectedImageState,
-    useEditorImageState,
-} from '@store/editorImageState';
-import { getAllArtworks } from '@utils/networking';
+import { useSelectedImageState, useEditorImageState } from '@store/editorImageState';
+import { getAllArtworks } from 'service/networking';
 import { Container, Tiles, Check } from './style';
 
 const Selector = () => {
@@ -20,11 +17,7 @@ const Selector = () => {
     const onClickImage = (id: number) => {
         return () => {
             const idx = selectedImages.findIndex((image) => image.id === id);
-            if (idx < 0)
-                setSelectedImages([
-                    ...selectedImages,
-                    images.find((image) => image.id === id)!,
-                ]);
+            if (idx < 0) setSelectedImages([...selectedImages, images.find((image) => image.id === id)!]);
             else {
                 const tmpSelectedImages = [...selectedImages];
                 tmpSelectedImages.splice(idx, 1);
@@ -38,9 +31,7 @@ const Selector = () => {
         <Container>
             <Tiles>
                 {images.map((image) => {
-                    const selected = selectedImages.findIndex(
-                        (img) => img.id === image.id,
-                    );
+                    const selected = selectedImages.findIndex((img) => img.id === image.id);
                     return (
                         <div key={image.id} onClick={onClickImage(image.id)}>
                             <img src={image.originalImage} alt={image.title} />
