+++ client/components/MyPage/AuctionPage/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import styled from '@emotion/styled';
 import Filter from '../Filter';
-import { getUserArtworkTrades } from '@utils/networking';
+import { getUserArtworkTrades } from 'service/networking';
 import { AuctionCardProps } from '@const/card-type';
 
 const filtering = ['거래', '입찰'];
